trailGenerate/requirements.txt
matplotlib==3.3.4
networkx==2.1
numba==0.47.0
numpy==1.18.0
opencv_python_headless==4.5.5.62
Pillow==8.4.0
scikit_image==0.17.2
sknw==0.14
llvmlite==0.31.0

